appview: rework RepoLanguages
- move endpoint on knot to /{did}/{name}/languages
- remove buggy percentage calculation
- fix endpoint to work on empty repos
- allowing calling endpoint with empty ref

+++ types/repo.go
@@ -112,5 +112,5 @@ type AncestorCheckResponse struct {
 
 type RepoLanguageResponse struct {
 	// Language: Percentage
-	Languages map[string]float64 `json:"languages"`
+	Languages map[string]int `json:"languages"`
 }
